refactor(inferencex_atom): update imports for inferencex_atom package paths
