+++ cvs/lib/inference/unittests/test_inferencex_atom_server_reuse.py
@@ -8,12 +8,12 @@ Unit tests for InferenceX ATOM server-reuse helpers and sweep parametrization.
 import unittest
 from types import SimpleNamespace
 
-from cvs.lib.inference.utils.inferencex_atom_config_loader import (
+from cvs.lib.inference.inferencex_atom.inferencex_atom_config_loader import (
     expand_sweep_parametrize,
     reuse_server_flag,
     server_session_key,
 )
-from cvs.lib.inference.utils.inferencex_atom_parsing import METRIC_TIER_ORDER
+from cvs.lib.inference.inferencex_atom.inferencex_atom_parsing import METRIC_TIER_ORDER
 
 
 class TestServerReuseHelpers(unittest.TestCase):
