+++ cvs/tests/inference/inferencex_atom/conftest.py
@@ -16,7 +16,7 @@ from cvs.lib.inference.utils.inference_suite_lifecycle import (
     html_metric_table_row,
     sort_lifecycle_items,
 )
-from cvs.lib.inference.utils.inferencex_atom_config_loader import (
+from cvs.lib.inference.inferencex_atom.inferencex_atom_config_loader import (
     load_variant,
     orchestrator_container_from_variant,
 )
